+++ packages/elementary_tools/README.md
@@ -1,2 +1,11 @@
 # Elementary tools
 Tools for better experience with Elementary package
+
+## [Elementary CLI](./elementary_cli)
+Utilities for generating/analyzing elementary-based code
+
+## [Intellij Plugin](./plugin_intellij)
+Plugin for intellij-based platforms
+
+## [VSCode Plugin](./plugin_vscode)
+Plugin for VSCode
