+++ packages/elementary_tools/plugin_vscode/.vscodeignore
@@ -8,3 +8,4 @@ vsc-extension-quickstart.md
 **/.eslintrc.json
 **/*.map
 **/*.ts
+.idea/**
