+++ packages/elementary_tools/plugin_vscode/CHANGELOG.md
@@ -4,10 +4,10 @@ All notable changes to the "elementary" extension will be documented in this fil
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [1.0.0]
+## [0.0.1]()
 
-- Added basic module generetion
+- Added basic module generation
 
-## [Unreleased]
+## [Unreleased]()
 
 - Initial release
