+++ packages/elementary_tools/plugin_vscode/README.md
@@ -1,74 +1,44 @@
-# elementary README
-
-## Dependencies installation
-`npm install git+https://github.com/Dart-Code/Dart-Code.git`
+# Elementary README
 
 <!-- TODO Rewrite -->
-This is the README for your extension "elementary". After writing up a brief description, we recommend including the following sections.
+Extension for better experience with Elementary
+
+## Dependencies installation
+* Dart-Code: `npm install git+https://github.com/Dart-Code/Dart-Code.git`
 
 ## Features
 
-Describe specific features of your extension including screenshots of your extension in action. Image paths are relative to this README file.
-
-For example if there is an image subfolder under your extension project workspace:
-
-\!\[feature X\]\(images/feature-x.png\)
-
-> Tip: Many popular extensions utilize animations. This is an excellent way to show off your extension! We recommend short, focused animations that are easy to follow.
+<!-- TODO add more features -->
+* Elementary module generator
 
 ## Requirements
 
-If you have any requirements or dependencies, add a section describing those and how to install and configure them.
+* Install [Elementary CLI][cli] to global pub
+
+[cli]: https://github.com/Elementary-team/flutter-elementary/tree/elementary_tools/packages/elementary_tools/elementary_cli
 
 ## Extension Settings
 
-Include if your extension adds any VS Code settings through the `contributes.configuration` extension point.
-
-For example:
-
 This extension contributes the following settings:
 
-* `myExtension.enable`: enable/disable this extension
-* `myExtension.thing`: set to `blah` to do something
+* `elementary.enable`: enable/disable this extension
+
+<!--
+* `elementary.testingDirectory`: set to testing directory path to
+generate tests there or leave empty to generate tests side-by-side
+with source files
+-->
 
 ## Known Issues
 
-Calling out known issues can help limit users opening duplicate issues against your extension.
+* Nothing serious for now ^-^
+* Elementary CLI is not automatically installed
 
 ## Release Notes
 
-Users appreciate release notes as you update your extension.
+* [See full changelog][changelog]
 
-### 1.0.0
+[changelog]: https://github.com/Elementary-team/flutter-elementary/tree/elementary_tools/packages/elementary_tools/plugin_vscode/CHANGELOG.md
+### 0.0.1-dev
 
-Initial release of ...
-
-### 1.0.1
-
-Fixed issue #.
-
-### 1.1.0
-
-Added features X, Y, and Z.
-
------------------------------------------------------------------------------------------------------------
-## Following extension guidelines
-
-Ensure that you've read through the extensions guidelines and follow the best practices for creating your extension.
-
-* [Extension Guidelines](https://code.visualstudio.com/api/references/extension-guidelines)
-
-## Working with Markdown
-
-**Note:** You can author your README using Visual Studio Code.  Here are some useful editor keyboard shortcuts:
-
-* Split the editor (`Cmd+\` on macOS or `Ctrl+\` on Windows and Linux)
-* Toggle preview (`Shift+CMD+V` on macOS or `Shift+Ctrl+V` on Windows and Linux)
-* Press `Ctrl+Space` (Windows, Linux) or `Cmd+Space` (macOS) to see a list of Markdown snippets
-
-### For more information
-
-* [Visual Studio Code's Markdown Support](http://code.visualstudio.com/docs/languages/markdown)
-* [Markdown Syntax Reference](https://help.github.com/articles/markdown-basics/)
-
-**Enjoy!**
+* Added Elementary module generation
